+++ scautable/js/src/excelStub.scala
@@ -1,0 +1,4 @@
+package io.github.quafadas.scautable
+
+
+object Excel
