+++ scautable/src/package.scala
@@ -5,5 +5,8 @@ object table:
   export io.github.quafadas.scautable.CSV
   export io.github.quafadas.scautable.CSV.*
   export io.github.quafadas.scautable.scautable
+  export io.github.quafadas.scautable.CsvIterator
+  export io.github.quafadas.scautable.Excel
+  export io.github.quafadas.scautable.Excel.*
   export io.github.quafadas.scautable.scautable.*
   export io.github.quafadas.scautable.ConsoleFormat.*
